requirements.txt
pandas
mlflow
notebook
numpy
scikit-learn
matplotlib
python-box
PyYAML
tqdm
ensure
joblib
types-pyYAML
Flask
Flask-Cors